Create csharp
We’re using the OpenAI Unity asset to generate a Unity C# script for a loot table. We’re using the ChatGPT tool to generate the script by sending a prompt to the OpenAI API. Once the script is generated, we’re saving it to a file in our Unity project.